requirements.txt
flask
python-docx
pdfplumber
groq
openpyxl
